+++ R/ci.R
@@ -85,7 +85,7 @@ ci.mlm <- function(x, ci = .95, ...) {
 #' @export
 ci.lm <- function(x, ci = .95, method = NULL, ...) {
   robust <- !is.null(method) && method == "robust"
-  ci_wald(model = x, ci = ci, component = "conditional", robust = robust, ...)
+  ci_wald(model = x, ci = ci, robust = robust, ...)
 }
 
 
@@ -145,9 +145,9 @@ ci.glm <- function(x, ci = .95, method = c("profile", "wald", "robust"), ...) {
     out <- lapply(ci, function(i) .ci_profiled(model = x, ci = i))
     out <- do.call(rbind, out)
   } else if (method == "robust") {
-    out <- ci_wald(model = x, ci = ci, component = "conditional", robust = TRUE, ...)
+    out <- ci_wald(model = x, ci = ci, robust = TRUE, ...)
   } else {
-    out <- ci_wald(model = x, ci = ci, component = "conditional")
+    out <- ci_wald(model = x, ci = ci)
   }
   row.names(out) <- NULL
   out
@@ -169,15 +169,15 @@ ci.polr <- function(x, ci = .95, method = c("profile", "wald", "robust"), ...) {
     out <- lapply(ci, function(i) .ci_profiled2(model = x, ci = i))
     out <- do.call(rbind, out)
   } else if (method == "robust") {
-    out <- ci_wald(model = x, ci = ci, component = "conditional", robust = TRUE, ...)
+    out <- ci_wald(model = x, ci = ci, robust = TRUE, ...)
   } else {
-    out <- ci_wald(model = x, ci = ci, component = "conditional")
+    out <- ci_wald(model = x, ci = ci)
   }
 
   # for polr, profiled CI do not return CI for response levels
   # thus, we also calculate Wald CI and add missing rows to result
 
-  out_missing <- ci_wald(model = x, ci = ci, component = "conditional")
+  out_missing <- ci_wald(model = x, ci = ci)
   missing_rows <- out_missing$Parameter %in% setdiff(out_missing$Parameter, out$Parameter)
   out <- rbind(out, out_missing[missing_rows, ])
 
@@ -285,6 +285,9 @@ ci.svyglm.zip <- ci.gamlss
 ci.vglm <- ci.gamlss
 
 #' @export
+ci.multinom <- ci.gamlss
+
+#' @export
 ci.svyglm.glimML <- ci.gamlss
 
 #' @export
@@ -385,7 +388,7 @@ ci.gls <- function(x, ci = .95, ...) {
 #' @export
 ci.lme <- function(x, ci = .95, ...) {
   if (!requireNamespace("nlme", quietly = TRUE)) {
-    ci_wald(model = x, ci = ci, component = "conditional")
+    ci_wald(model = x, ci = ci)
   } else {
     out <- lapply(ci, function(i) {
       ci_list <- nlme::intervals(x, level = i, ...)
@@ -404,7 +407,7 @@ ci.lme <- function(x, ci = .95, ...) {
 #' @importFrom insight print_color
 #' @importFrom stats qnorm
 #' @export
-ci.parameters_std_classic <- function(x, ci = .95, ...) {
+ci.effectsize_std_params <- function(x, ci = .95, ...) {
   se <- attr(x, "standard_error")
 
   if (is.null(se)) {