+++ R/dof.R
@@ -37,7 +37,7 @@ degrees_of_freedom <- function(model, method = "analytical") {
 
   if (method == "any") {
     dof <- .degrees_of_freedom_fit(model, verbose = FALSE)
-    if (is.null(dof)) {
+    if (is.null(dof) || is.infinite(dof)) {
       dof <- .degrees_of_freedom_analytical(model, kenward = FALSE)
     }
   } else if (method == "analytical") {
@@ -106,7 +106,7 @@ dof <- degrees_of_freedom
 
   # 2nd try
   if (inherits(dof, "try-error")) {
-    dof <- NULL
+    dof <- Inf
     if (verbose) {
       insight::print_color("Could not extract degrees of freedom.\n", "red")
     }