+++ R/model_parameters.gam.R
@@ -3,7 +3,7 @@
 #' Extract and compute indices and measures to describe parameters of generalized additive models (GAM(M)s).
 #'
 #' @param model A gam/gamm model.
-#' @inheritParams model_parameters.lm
+#' @inheritParams model_parameters.default
 #'
 #' @seealso \code{\link[=standardize_names]{standardize_names()}} to rename
 #'   columns into a consistent, standardized naming scheme.
@@ -26,8 +26,7 @@ model_parameters.gam <- function(model, ci = .95, bootstrap = FALSE, iterations 
     parameters <- .extract_parameters_generic(model, ci = ci, component = "all", merge_by = c("Parameter", "Component"))
   }
 
-  attr(parameters, "pretty_names") <- format_parameters(model)
-  attr(parameters, "ci") <- ci
+  parameters <- .add_model_parameters_attributes(parameters, model, ci, ...)
   class(parameters) <- c("parameters_model", "see_parameters_model", class(parameters))
   parameters
 }