+++ R/model_parameters.lavaan.R
@@ -4,7 +4,7 @@
 #'
 #' @param model CFA or SEM created by the \code{lavaan::cfa} or \code{lavaan::sem} functions.
 #' @param standardize Return standardized parameters (standardized coefficients). See \code{lavaan::standardizedsolution}.
-#' @inheritParams model_parameters.lm
+#' @inheritParams model_parameters.default
 #' @param type What type of links to return. Can be \code{"all"} or some of \code{c("regression", "correlation", "loading", "variance", "mean")}.
 #' @param ... Arguments passed to or from other methods.
 #'