+++ R/model_parameters.wbm.R
@@ -1,5 +1,5 @@
 #' @export
-model_parameters.wbm <- function(model, ci = .95, bootstrap = FALSE, ci_method = "wald", iterations = 1000, ...) {
+model_parameters.wbm <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, ...) {
   # Processing
   if (bootstrap) {
     parameters <- parameters_bootstrap(model, iterations = iterations, ci = ci, ...)
@@ -7,8 +7,7 @@ model_parameters.wbm <- function(model, ci = .95, bootstrap = FALSE, ci_method =
     parameters <- .extract_parameters_generic(model, ci = ci, component = "conditional", merge_by = c("Parameter", "Component"))
   }
 
-  attr(parameters, "pretty_names") <- format_parameters(model)
-  attr(parameters, "ci") <- ci
+  parameters <- .add_model_parameters_attributes(parameters, model, ci, ...)
   class(parameters) <- c("parameters_model", "see_parameters_model", class(parameters))
   parameters
 }