+++ R/model_parameters_mixed.R
@@ -1,11 +1,11 @@
 #' Mixed Model Parameters
 #'
-#' Parameters of mixed models.
+#' Parameters of (linear) mixed models.
 #'
 #' @param model A mixed model.
-#' @inheritParams model_parameters.lm
+#' @inheritParams model_parameters.default
 #' @param p_method Method for computing p values. See \code{\link[=p_value]{p_value()}}.
-#' @param ci_method Method for computing confidence intervals (CI). See \code{\link[=ci]{ci()}}.
+#' @param ci_method Method for computing confidence intervals (CI). See \code{\link[=ci.merMod]{ci()}}.
 #'
 #' @seealso \code{\link[=standardize_names]{standardize_names()}} to rename
 #'   columns into a consistent, standardized naming scheme.
@@ -42,9 +42,39 @@ model_parameters.merMod <- function(model, ci = .95, bootstrap = FALSE, p_method
   }
 
 
-
-  attr(parameters, "pretty_names") <- format_parameters(model)
-  attr(parameters, "ci") <- ci
+  parameters <- .add_model_parameters_attributes(parameters, model, ci, ...)
   class(parameters) <- c("parameters_model", "see_parameters_model", class(parameters))
   parameters
 }
+
+
+
+
+# Mixed Models with zero inflation ------------------------------------
+
+#' @inheritParams model_simulate
+#' @rdname model_parameters.merMod
+#' @export
+model_parameters.glmmTMB <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, component = c("all", "conditional", "zi", "zero_inflated"), ...) {
+  component <- match.arg(component)
+
+  # fix argument, if model has no zi-part
+  if (!insight::model_info(model)$is_zero_inflated && component != "conditional") {
+    component <- "conditional"
+  }
+
+  # Processing
+  if (bootstrap) {
+    parameters <- parameters_bootstrap(model, iterations = iterations, ci = ci, ...)
+  } else {
+    parameters <- .extract_parameters_generic(model, ci = ci, component = component, ...)
+  }
+
+
+  parameters <- .add_model_parameters_attributes(parameters, model, ci, ...)
+  class(parameters) <- c("parameters_model", "see_parameters_model", class(parameters))
+  parameters
+}
+
+#' @export
+model_parameters.MixMod <- model_parameters.glmmTMB